posts/colpali/requirements.txt
python-fasthtml
modal
notebook
pytest
pytest-xdist
torch
marko
pillow